Added randomization in populateTiles
Added randomization by creating a Random instance before the loop that adds the Tiles to the panel. The randomization is done by getting a random number from the number of tiles that have not yet been placed out. Technically this is done by calling random.nextInt and passing the size of the tile list to get an index. The Tile with the corresponding index is then retrieved from the list and removed from the list by using the remove(i) call.